Move revisions card components to one file
