--- client/src/pages/ProfilePage/components/RevisionsCard/RevisionsCard.jsx
@@ -1,37 +1,0 @@
-import { useState, useEffect } from "react";
-import propTypes from "prop-types";
-import RevisionsGrid from "./RevisionsGrid";
-
-RevisionsCard.propTypes = {
-    userInfo: propTypes.object.isRequired,
-    JWT: propTypes.string.isRequired,
-    JWTSetter: propTypes.func.isRequired,
-};
-
-function RevisionsCard(props) {
-    const [problemsToRevise, setProblemsToRevise] = useState([]);
-
-    useEffect(() => {
-        if (props.userInfo.problemStatuses === null || props.userInfo.problemStatuses === undefined) {
-            return;
-        }
-        setProblemsToRevise(props.userInfo.problemStatuses.filter((status) => status.AC === 0 || status.markedForRevision));
-    }, [props.userInfo]);
-
-    return (
-        <div className="card shadow m-4">
-            <div className="card-header" data-bs-toggle="collapse" data-bs-target="#revisions-body" role="button">
-                <b>Revisions</b>
-            </div>
-            <div className="collapse show" id="revisions-body">
-                <div className="container-fluid">
-                    <div className="card-body">
-                        <RevisionsGrid problemStatuses={problemsToRevise} JWT={props.JWT} JWTSetter={props.JWTSetter} />
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-}
-
-export default RevisionsCard;
